Some tweaks and add doxyegn comments
