+++ examples/UnitUnified/UnitCO2/PlotToSerial/main/PlotToSerial.cpp
@@ -6,27 +6,15 @@
 /*
   Example using M5UnitUnified for UnitCO2
 */
-
-// #define USING_PAHUB (2)  // Connection channel number for use via PaHub.
-// #define USING_M5HAL      // When using M5HAL
-
+//#define USING_M5HAL  // When using M5HAL
 #include <M5Unified.h>
 #include <M5UnitUnified.h>
 #include <M5UnitUnifiedENV.h>
-#if defined(USING_PAHUB)
-#include <M5UnitUnifiedHUB.h>
-#endif
 
 namespace {
 auto& lcd = M5.Display;
-
 m5::unit::UnitUnified Units;
 m5::unit::UnitCO2 unit;
-
-#if defined(USING_PAHUB)
-m5::unit::UnitPaHub2 unitPaHub;
-#endif
-
 }  // namespace
 
 void setup() {
@@ -36,47 +24,6 @@ void setup() {
     auto pin_num_scl = M5.getPin(m5::pin_name_t::port_a_scl);
     M5_LOGI("getPin: SDA:%u SCL:%u", pin_num_sda, pin_num_scl);
 
-#if 0
-    auto cfg           = unit.config();
-    cfg.start_periodic = false;
-    unit.config(cfg);
-#endif
-
-#if defined(USING_PAHUB)
-#pragma message "Using via PaHub"
-    // Using via PaHub
-#if defined(USING_M5HAL)
-#pragma message "Using M5HAL"
-    // Using M5HAL
-    m5::hal::bus::I2CBusConfig i2c_cfg;
-    i2c_cfg.pin_sda = m5::hal::gpio::getPin(pin_num_sda);
-    i2c_cfg.pin_scl = m5::hal::gpio::getPin(pin_num_scl);
-    auto i2c_bus    = m5::hal::bus::i2c::getBus(i2c_cfg);
-
-    if (!unitPaHub.add(unit, USING_PAHUB) || !Units.add(unitPaHub, i2c_bus ? i2c_bus.value() : nullptr) ||
-        !Units.begin()) {
-        M5_LOGE("Failed to begin");
-        lcd.clear(TFT_RED);
-        while (true) {
-            m5::utility::delay(10000);
-        }
-    }
-#else
-    // Using TwoWire
-#pragma message "Using Wire"
-    Wire.begin(pin_num_sda, pin_num_scl, 400000U);
-    if (!unitPaHub.add(unit, USING_PAHUB) || !Units.add(unitPaHub, Wire) || !Units.begin()) {
-        M5_LOGE("Failed to begin");
-        lcd.clear(TFT_RED);
-        while (true) {
-            m5::utility::delay(10000);
-        }
-    }
-#endif
-
-#else
-    // Direct connection
-#pragma message "Direct connection"
 #if defined(USING_M5HAL)
 #pragma message "Using M5HAL"
     // Using M5HAL
@@ -103,20 +50,6 @@ void setup() {
         }
     }
 #endif
-#endif
-
-#if 0    
-    // This is the imagined code that will happen after M5HAL is serviced and
-    // M5Unified is modified accordingly.
-    if (!M5.Units.add(unit, M5.PortA) || !M5.Units.begin()) {
-        M5_LOGE("Failed to begin");
-        lcd.clear(TFT_RED);
-        while (true) {
-            m5::utility::delay(10000);
-        }
-    }
-#endif
-
     M5_LOGI("M5UnitUnified has been begun");
     M5_LOGI("%s", Units.debugInfo().c_str());
 
@@ -126,13 +59,6 @@ void setup() {
 void loop() {
     M5.update();
     Units.update();
-
-#if 0    
-    // This is the imagined code that will happen after M5HAL is serviced and
-    // M5Unified is modified accordingly.
-    M5.update(); // Call M5.Units.update() in it.
-#endif
-
     if (unit.updated()) {
         // Can be checked e.g. by serial plotters
         M5_LOGI("\n>CO2:%u\n>Temperature:%2.2f\n>Humidity:%2.2f", unit.co2(), unit.temperature(), unit.humidity());
