+++ src/unit/unit_SCD40.hpp
@@ -11,12 +11,64 @@
 #define M5_UNIT_ENV_UNIT_SCD40_HPP
 
 #include <M5UnitComponent.hpp>
-#include "internal/scd4x.hpp"
 #include <m5_utility/container/circular_buffer.hpp>
 #include <limits>  // NaN
 
 namespace m5 {
 namespace unit {
+
+/*!
+  @namespace scd4x
+  @brief For SCD40/41
+ */
+namespace scd4x {
+/*!
+  @enum Mode
+  @brief Mode of periodic measurement
+ */
+enum class Mode : uint8_t {
+    Normal,    //!< Normal (Receive data every 5 seconds)
+    LowPower,  //!< Low power (Receive data every 30 seconds)
+};
+
+/*!
+  @struct Data
+  @brief Measurement data group
+ */
+struct Data {
+    std::array<uint8_t, 9> raw{};  //!< @brief RAW data
+    uint16_t co2() const;          //!< @brief CO2 concentration (ppm)
+    //! @brief temperature (Celsius)
+    inline float temperature() const {
+        return celsius();
+    }
+    float celsius() const;     //!< @brief temperature (Celsius)
+    float fahrenheit() const;  //!< @brief temperature (Fahrenheit)
+    float humidity() const;    //!< @brief humidity (RH)
+};
+
+///@cond
+// Max command duration(ms)
+// For SCD40/41
+constexpr uint16_t READ_MEASUREMENT_DURATION{1};
+constexpr uint16_t STOP_PERIODIC_MEASUREMENT_DURATION{500};
+constexpr uint16_t SET_TEMPERATURE_OFFSET_DURATION{1};
+constexpr uint16_t GET_TEMPERATURE_OFFSET_DURATION{1};
+constexpr uint16_t SET_SENSOR_ALTITUDE_DURATION{1};
+constexpr uint16_t GET_SENSOR_ALTITUDE_DURATION{1};
+constexpr uint16_t SET_AMBIENT_PRESSURE_DURATION{1};
+constexpr uint16_t PERFORM_FORCED_CALIBRATION_DURATION{400};
+constexpr uint16_t SET_AUTOMATIC_SELF_CALIBRATION_ENABLED_DURATION{1};
+constexpr uint16_t GET_AUTOMATIC_SELF_CALIBRATION_ENABLED_DURATION{1};
+constexpr uint16_t GET_DATA_READY_STATUS_DURATION{1};
+constexpr uint16_t PERSIST_SETTINGS_DURATION{800};
+constexpr uint16_t GET_SERIAL_NUMBER_DURATION{1};
+constexpr uint16_t PERFORM_SELF_TEST_DURATION{10000};
+constexpr uint16_t PERFORM_FACTORY_RESET_DURATION{1200};
+constexpr uint16_t REINIT_DURATION{20};
+/// @endcond
+
+}  // namespace scd4x
 
 /*!
   @class UnitSCD40
@@ -90,11 +142,27 @@ class UnitSCD40 : public Component, public PeriodicMeasurementAdapter<UnitSCD40,
     ///@name Periodic measurement
     ///@{
     /*!
-      @brief Start low power periodic measurements
+      @brief Start periodic measurement
+      @param mode Measurement mode
+      @return True if successful
+    */
+    inline bool startPeriodicMeasurement(const scd4x::Mode mode = scd4x::Mode::Normal) {
+        return PeriodicMeasurementAdapter<UnitSCD40, scd4x::Data>::startPeriodicMeasurement(mode);
+    }
+    /*!
+      @brief Start low power periodic measurement
       @return True if successful
     */
     inline bool startLowPowerPeriodicMeasurement() {
         return startPeriodicMeasurement(scd4x::Mode::LowPower);
+    }
+    /*!
+      @brief Stop periodic measurement
+      @param duration Max command duration(ms)
+      @return True if successful
+    */
+    inline bool stopPeriodicMeasurement(const uint32_t duration = scd4x::STOP_PERIODIC_MEASUREMENT_DURATION) {
+        return PeriodicMeasurementAdapter<UnitSCD40, scd4x::Data>::stopPeriodicMeasurement(duration);
     }
     ///@}
 
@@ -226,25 +294,8 @@ class UnitSCD40 : public Component, public PeriodicMeasurementAdapter<UnitSCD40,
     ///@}
 
    protected:
-    ///@note Call via startPeriodicMeasurement/stopPeriodicMeasurement
-    ///@name Periodic measurement
-    ///@{
-    /*!
-      @brief Start periodic measurement
-      @param mode Mode of periodic measurement
-      @return True if successful
-    */
     bool start_periodic_measurement(const scd4x::Mode mode = scd4x::Mode::Normal);
-    /*!
-      @brief Stop measurement
-      @param duration Max command duration(ms)
-      @warning The sensor will only respond to other commands after waiting 500
-      ms after issuing the stop_periodic_measurement command
-      @return True if successful
-    */
     bool stop_periodic_measurement(const uint32_t duration = scd4x::STOP_PERIODIC_MEASUREMENT_DURATION);
-    ///@}
-
     bool read_data_ready_status();
     bool read_measurement(scd4x::Data &d, const bool all = true);
 
@@ -254,6 +305,36 @@ class UnitSCD40 : public Component, public PeriodicMeasurementAdapter<UnitSCD40,
     std::unique_ptr<m5::container::CircularBuffer<scd4x::Data>> _data{};
     config_t _cfg{};
 };
+
+///@cond
+namespace scd4x {
+namespace command {
+// Basic Commands
+constexpr uint16_t START_PERIODIC_MEASUREMENT{0x21b1};
+constexpr uint16_t READ_MEASUREMENT{0xec05};
+constexpr uint16_t STOP_PERIODIC_MEASUREMENT{0x3f86};
+// On-chip output signal compensation
+constexpr uint16_t SET_TEMPERATURE_OFFSET{0x241d};
+constexpr uint16_t GET_TEMPERATURE_OFFSET{0x2318};
+constexpr uint16_t SET_SENSOR_ALTITUDE{0x2427};
+constexpr uint16_t GET_SENSOR_ALTITUDE{0x2322};
+constexpr uint16_t SET_AMBIENT_PRESSURE{0xe000};
+// Field calibration
+constexpr uint16_t PERFORM_FORCED_CALIBRATION{0x362f};
+constexpr uint16_t SET_AUTOMATIC_SELF_CALIBRATION_ENABLED{0x2416};
+constexpr uint16_t GET_AUTOMATIC_SELF_CALIBRATION_ENABLED{0x2313};
+// Low power
+constexpr uint16_t START_LOW_POWER_PERIODIC_MEASUREMENT{0x21ac};
+constexpr uint16_t GET_DATA_READY_STATUS{0xe4b8};
+// Advanced features
+constexpr uint16_t PERSIST_SETTINGS{0x3615};
+constexpr uint16_t GET_SERIAL_NUMBER{0x3682};
+constexpr uint16_t PERFORM_SELF_TEST{0x3639};
+constexpr uint16_t PERFORM_FACTORY_RESET{0x3632};
+constexpr uint16_t REINIT{0x3646};
+}  // namespace command
+}  // namespace scd4x
+///@endcond
 
 }  // namespace unit
 }  // namespace m5
